chore: merge utils
